requirements.txt
## minimum requirements.txt
streamlit>=1.32
langchain>=0.3
langchain-openai>=0.3
httpx>=0.27
python-dotenv>=1.0